% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_calculator.R
\name{get_binom_prob}
\alias{get_binom_prob}
\title{Combine all probabilities to get our prob density}
\usage{
get_binom_prob(prob_hit, prob_wound, prob_go_trough_save)
}
\arguments{
\item{hit}{Prob to hit, prob to wound and the pro to go through armor}
}
\value{
combined probability
}
\description{
Combine all probabilities to get our prob density
}
