% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_calculator.R
\name{get_kill_count}
\alias{get_kill_count}
\title{Get the number of dead units knowing how many wounds have been dealt, with how many damages per shot and the target's wound characteristic}
\usage{
get_kill_count(nbwounds, damage, lifepoints)
}
\arguments{
\item{nbwounds}{Total number of wounds dealt after hitting, wounding and armorsaving}

\item{damage}{Damage profile of the weapon}

\item{lifepoints}{Total wounds of the model}
}
\value{
Number of dead model
}
\description{
Get the number of dead units knowing how many wounds have been dealt, with how many damages per shot and the target's wound characteristic
}
\examples{
get_kill_count(4)
get_kill_count(2)
}
