% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_calculator.R
\name{get_prob_hit}
\alias{get_prob_hit}
\title{Get the probability needed to hit}
\usage{
get_prob_hit(hit)
}
\arguments{
\item{hit}{Ballistic skill or Weapon skill. It's a roll needed}
}
\value{
Probability to hit
}
\description{
Get the probability needed to hit
}
\examples{
get_prob_hit(5)
get_prob_hit(2)
}
