% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_calculator.R
\name{get_prob_save}
\alias{get_prob_save}
\title{Get the probability for a wound to go througth knowing the save of the opponent}
\usage{
get_prob_save(armorsave)
}
\arguments{
\item{armorsave}{Armor save in the datasheet}
}
\value{
Probability for a wound to go through knowing the target armor save
}
\description{
Get the probability for a wound to go througth knowing the save of the opponent
}
\examples{
get_prob_save(4)
get_prob_save(2)
}
