% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_calculator.R
\name{get_prob_wound}
\alias{get_prob_wound}
\title{Get the probability needed to wound}
\usage{
get_prob_wound(strength, toughness)
}
\arguments{
\item{strength}{Weapon strength.}

\item{toughness}{Toughness of the target.}
}
\value{
Probability to wound
}
\description{
Get the probability needed to wound
}
\examples{
get_prob_wound(4, 4)
get_prob_wound(4, 8)
}
