% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_calculator.R
\name{get_strength_toughness_prob}
\alias{get_strength_toughness_prob}
\title{Get the roll needed to wound}
\usage{
get_strength_toughness_prob(ST, TO)
}
\arguments{
\item{ST}{Weapon strength.}

\item{TO}{Toughness of the target.}
}
\value{
Needed wound roll
}
\description{
Get the roll needed to wound
}
\examples{
get_strength_toughness_prob(4, 4)
get_strength_toughness_prob(10,5)
}
