% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_calculator.R
\name{get_wound_count}
\alias{get_wound_count}
\title{Get the number of wounds possible using the damages per shot.}
\usage{
get_wound_count(nbwounds, damage)
}
\arguments{
\item{nbwounds}{Total number of wounds dealt after hitting, wounding and armorsaving}

\item{damage}{Damage profile of the weapon}
}
\value{
Number of dead model
}
\description{
Get the number of wounds possible using the damages per shot.
}
\examples{
get_prob_save(4)
get_prob_save(2)
}
